move navigation graph out to different file
