+++ app/src/main/java/com/cattailsw/mediaplayer/ExoPlayerScreen.kt
@@ -1,0 +1,57 @@
+package com.cattailsw.mediaplayer
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.Icon
+import androidx.compose.material.IconButton
+import androidx.compose.material.Surface
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.ArrowBack
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.DisposableEffect
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.viewinterop.AndroidView
+import com.cattailsw.mediaplayer.ui.theme.CMediaPlayerTheme
+import com.google.android.exoplayer2.Player
+import com.google.android.exoplayer2.ui.StyledPlayerView
+
+@Composable
+fun ExoPlayerScreen(
+    player: Player,
+    modifier: Modifier = Modifier,
+    onBack: () -> Unit = {},
+    onDisposeAction: () -> Unit = {},
+) {
+    val context = LocalContext.current
+
+    val playWhenReady by rememberSaveable { mutableStateOf(true) }
+
+    player.playWhenReady = playWhenReady
+
+    CMediaPlayerTheme(darkTheme = true) {
+        Surface {
+            Box(modifier = modifier.fillMaxSize()) {
+                DisposableEffect(key1 = Unit, effect = {
+                    onDispose {
+                        onDisposeAction()
+                    }
+                })
+                AndroidView(
+                    factory = {
+                        StyledPlayerView(context).apply {
+                            this.player = player
+                        }
+                    },
+                    modifier = Modifier.fillMaxSize()
+                )
+                IconButton(onClick = onBack) {
+                    Icon(imageVector = Icons.Filled.ArrowBack, contentDescription = "Back")
+                }
+            }
+        }
+    }
+}
